modif minime

+++ libBigInt/src/entier.cpp
@@ -288,7 +288,7 @@ Entier operator%(const Entier& a, const Entier& b)
 
 ostream& operator<<(ostream& flux, const Entier& val)
 {
-    for(int i = val.valeur.size()-1; i > 0; i--)
+    for(size_t i = val.valeur.size()-1; i > 0; i--)
         flux << val.valeur[i] << '.';
     flux << val.valeur[0];
